+++ chat/src/main/java/ru/job4j/chat/controller/RoleController.java
@@ -3,6 +3,7 @@ package ru.job4j.chat.controller;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import ru.job4j.chat.model.Role;
 import ru.job4j.chat.repository.RoleRepository;
 
@@ -23,14 +24,18 @@ public class RoleController {
     @GetMapping("/{id}")
     public ResponseEntity<Role> findById(@PathVariable int id) {
         var role = this.repository.findById(id);
-        return new ResponseEntity<Role>(
-                role.orElse(new Role()),
-                role.isPresent() ? HttpStatus.OK : HttpStatus.NOT_FOUND
-        );
+        if (!role.isPresent()) {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, "Role is not found. Please, check requisites.");
+        }
+        return new ResponseEntity<Role>(role.orElse(new Role()), HttpStatus.OK);
     }
 
     @PostMapping("/")
     public ResponseEntity<Role> create(@RequestBody Role role) {
+        if (role.getAuthority() == null) {
+            throw new NullPointerException("Authority mustn't be empty");
+        }
         return new ResponseEntity<Role>(
                 this.repository.save(role),
                 HttpStatus.CREATED
@@ -39,12 +44,18 @@ public class RoleController {
 
     @PutMapping("/")
     public ResponseEntity<Void> update(@RequestBody Role role) {
+        if (role.getAuthority() == null) {
+            throw new NullPointerException("Authority mustn't be empty");
+        }
         this.repository.save(role);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable int id) {
+        if (id == 0) {
+            throw new NullPointerException("ID mustn't be empty");
+        }
         Role role = new Role();
         role.setId(id);
         this.repository.delete(role);
