+++ chat/src/main/java/ru/job4j/chat/controller/RoomController.java
@@ -3,6 +3,7 @@ package ru.job4j.chat.controller;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import ru.job4j.chat.model.Room;
 import ru.job4j.chat.service.RoomService;
 
@@ -23,14 +24,18 @@ public class RoomController {
     @GetMapping("/{id}")
     public ResponseEntity<Room> findById(@PathVariable int id) {
         var person = this.service.findById(id);
-        return new ResponseEntity<Room>(
-                person.orElse(new Room()),
-                person.isPresent() ? HttpStatus.OK : HttpStatus.NOT_FOUND
-        );
+        if (!person.isPresent()) {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, "Person is not found. Please, check requisites.");
+        }
+        return new ResponseEntity<Room>(person.orElse(new Room()), HttpStatus.OK);
     }
 
     @PostMapping("/")
     public ResponseEntity<Room> create(@RequestBody Room room) {
+        if (room.getName() == null) {
+            throw new NullPointerException("Username mustn't be empty");
+        }
         return new ResponseEntity<Room>(
                 this.service.save(room),
                 HttpStatus.CREATED
@@ -39,12 +44,18 @@ public class RoomController {
 
     @PutMapping("/")
     public ResponseEntity<Void> update(@RequestBody Room room) {
+        if (room.getName() == null) {
+            throw new NullPointerException("Username mustn't be empty");
+        }
         this.service.save(room);
         return ResponseEntity.ok().build();
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete(@PathVariable int id) {
+        if (id == 0) {
+            throw new NullPointerException("ID mustn't be empty");
+        }
         Room room = new Room();
         room.setId(id);
         this.service.delete(room);
